Extract merge.json grid parser into vc_merge_grid static lib
gmResolveGrid + gmCheckConnected and the GMSurfaceSpec / GMEdgeSpec
PODs are moved verbatim into apps/src/vc_merge_tifxyz_grid.{hpp,cpp}
and exposed under namespace vc::merge. The apps/CMakeLists.txt grows
a vc_merge_grid STATIC library that vc_merge_tifxyz now links against
(replacing the direct vc_core link); follow-up commits add a doctest
that links the same lib for parser-level regression coverage.

No behavior change: vc_merge_tifxyz --help and full-run paths are
unchanged.
